Fix the structure of view models
