+++ src/Tweb.Modules/eTweb.Application/Catalog/Products/IPubishProductService.cs
@@ -1,6 +1,7 @@
-﻿using eTweb.Application.Catalog.Dtos;
-using eTweb.Application.Catalog.Dtos.Public;
+﻿
 using eTweb.Application.Dtos;
+using eTweb.ViewModels.Catalog.Products;
+using eTweb.ViewModels.Catalog.Products.Public;
 using System;
 using System.Collections.Generic;
 using System.Text;
