+++ src/Tweb.Modules/eTweb.ViewModels/Catalog/Products/Manage/ProductCreateRequest.cs
@@ -2,11 +2,18 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace eTweb.Application.Catalog.Dtos.Manage
+namespace eTweb.ViewModels.Catalog.Products.Manage
 {
-    public class ProductUpdateRequest
+    public class ProductCreateRequest
     {
-        public int Id { get; set; }
+        public int Id { set; get; }
+        public decimal Price { set; get; }
+        public decimal OriginalPrice { set; get; }
+        public int Stock { set; get; }
+        public int ViewCount { set; get; }
+        public DateTime DateCreated { set; get; }
+
+        public bool? IsFeatured { get; set; }
         public string Name { set; get; }
         public string Description { set; get; }
         public string Details { set; get; }
