+++ src/Tweb.Modules/eTweb.ViewModels/Catalog/Products/ProductViewModel.cs
@@ -2,9 +2,9 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace eTweb.Application.Catalog.Dtos.Manage
-{ 
-    public class ProductCreateRequest
+namespace eTweb.ViewModels.Catalog.Products
+{
+    public class ProductViewModel
     {
         public int Id { set; get; }
         public decimal Price { set; get; }
@@ -12,8 +12,6 @@ namespace eTweb.Application.Catalog.Dtos.Manage
         public int Stock { set; get; }
         public int ViewCount { set; get; }
         public DateTime DateCreated { set; get; }
-
-        public bool? IsFeatured { get; set; }
         public string Name { set; get; }
         public string Description { set; get; }
         public string Details { set; get; }
